Update oai backend
A fundamental problem with zbMATH and swMATH identifiers is that they don't have a prefix. Therefore the full identifier is not part of the plain rest API result. This caused that a check in the OAI-PMH API failed. To disable this check a new switch has been introduced in version 1.2.11.